Created unit tests for scenarios 5 & 6
These unit tests simulate and evaluate the results of the following user activities:
User generates a password and chooses to copy the password to the clipboard. User quits.
User generates a password and chooses to save it to the associated account. User quits.

A 'conftest.py' file is also added for teat configuration purposes.