Create app.py
This code runs the create.html file and lets the user choose the specifications for the password and then type the name of the website that this password is going to be associated with. Then it renders the page result.html which shows the associated website, the generated password and the password strength bar.