Added functions to database.py and create_read.py
Added a new function to database.py that returns a list of all created users.
Added a new function to create_read.py that pulls new_user.html whenever the user attempts to create a new user. If the provided username already exists within the database, the webpage states as such and instructs the user to reattempt creating an account. If the provided username does not already exist, the server adds an entry to the database containing the provided username and a hash of the provided password. The webpage then welcomes the user and instructs them to return to the login page.