Added login functionality to create_read.py
Adjusted create_read.py such that when the user visits the site and attempts to login, the program will poll the database for the inputted login info. If matching info is found, the site will take the user to the page that is meant to display their saved passwords. If matching info is not found, the site will redirect the user back to the login page and display an error message.